requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pyyaml>=6.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
pytest>=7.4.0
pytest-timeout>=2.1.0
